pyproject.toml
[tool.poetry]
name = "shop-mts-by"
version = "0.1.0"
description = ""
authors = ["Vladislav.H <vkhrabrov94@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
pytest = "6.2.5"
selenium = "^4.1.0"
requests = "^2.27.1"
allure-pytest = "2.9.45"
pytest-rerunfailures = "^10.2"
pytest-xdist = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
